requirements.txt
asgiref==3.5.2
Django==4.1.4
django-crispy-forms==1.14.0
django-environ==0.9.0
django-ipware==4.0.2
gunicorn==20.1.0
mysqlclient==2.1.1
sqlparse==0.4.3
tzdata==2022.7
